clone button actually working
